git-svn-id: https://mattari.mydns.jp/repos/android_apps/trunk/NitechBBSrv@505 ebda3f91-3b32-42d3-bb0a-b50f2d3b12aa

+++ app/models/service/BBItemHead/BBItemHeadModelService.java
@@ -51,7 +51,9 @@ public class BBItemHeadModelService implements ModelService<Long, BBItemHead> {
 
 	@Override
 	public void delete(BBItemHead entry) {
-		entry.delete();
+		if (entry != null) {
+			entry.delete();
+		}
 	}
